refactor(web): shared SSRF guard for fetch_url and web.run
Extract the DNS-pin + restricted-IP SSRF preflight from fetch_url into
tools::web::guard and route web_run open/click fetches through the same
path so private, loopback, metadata, and unsafe redirect targets are
refused with consistent permission errors.
