+++ crates/tui/src/tools/web/mod.rs
@@ -1,0 +1,3 @@
+//! Shared web-tool helpers (SSRF guard, etc.).
+
+pub(crate) mod guard;
